Add helper folder and fix references
